requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==1.10.26
python-multipart==0.0.6
python-dotenv==1.0.0

# Neo4j Graph Database
neo4j==5.16.0

# NLP & Entity Extraction
spacy==3.8.11

# HTTP Client
requests==2.31.0

# Google Gemini SDK
google-generativeai==0.8.3

# Frontend
streamlit==1.30.0
pandas==2.2.0

# Note: After installing, download spaCy language model with:
# python -m spacy download en_core_web_sm
# Additional dependencies can be added as needed